Struct rustcoalescence_algorithms_cuda::kernel::launch::with_dimensions::SimulationKernelWithDimensions

pub struct SimulationKernelWithDimensions<'k, H: Habitat + RustToCuda, G: PrimeableRng + RustToCuda, R: LineageReference<H> + DeviceCopy, S: LineageStore<H, R> + RustToCuda, X: EmigrationExit<H, G, R, S> + RustToCuda, D: DispersalSampler<H, G> + RustToCuda, C: CoalescenceSampler<H, R, S> + RustToCuda, T: TurnoverRate<H> + RustToCuda, N: SpeciationProbability<H> + RustToCuda, E: MinSpeciationTrackingEventSampler<H, G, R, S, X, D, C, T, N> + RustToCuda, I: ImmigrationEntry + RustToCuda, A: SingularActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> + PeekableActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> + RustToCuda, ReportSpeciation: Boolean, ReportDispersal: Boolean> {
    pub(super) compiler: &'k mut PtxJITCompiler,
    pub(super) ptx_jit: bool,
    pub(super) module: &'k mut Module,
    pub(super) entry_point: &'k mut Function<'k>,
    pub(super) marker: PhantomData<(H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal)>,
    pub(super) grid_size: GridSize,
    pub(super) block_size: BlockSize,
    pub(super) shared_mem_bytes: u32,
}

Fields
compiler: &'k mut PtxJITCompilerptx_jit: boolmodule: &'k mut Moduleentry_point: &'k mut Function<'k>marker: PhantomData<(H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal)>grid_size: GridSizeblock_size: BlockSizeshared_mem_bytes: u32
Implementations

impl<'k, H: Habitat + RustToCuda, G: PrimeableRng + RustToCuda, R: LineageReference<H> + DeviceCopy, S: LineageStore<H, R> + RustToCuda, X: EmigrationExit<H, G, R, S> + RustToCuda, D: DispersalSampler<H, G> + RustToCuda, C: CoalescenceSampler<H, R, S> + RustToCuda, T: TurnoverRate<H> + RustToCuda, N: SpeciationProbability<H> + RustToCuda, E: MinSpeciationTrackingEventSampler<H, G, R, S, X, D, C, T, N> + RustToCuda, I: ImmigrationEntry + RustToCuda, A: SingularActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> + PeekableActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> + RustToCuda, ReportSpeciation: Boolean, ReportDispersal: Boolean> SimulationKernelWithDimensions<'k, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>

pub fn with_stream<'s>(
self,
stream: &'s Stream
) -> SimulationKernelWithDimensionsStream<'k, 's, H, G, R, S, X, D, C, T, N, E, I, A, ReportSpeciation, ReportDispersal>