Function rustcoalescence_algorithms_cuda::kernel::specialiser::get_ptx_cstr

pub fn get_ptx_cstr<H: Habitat + RustToCuda, G: PrimeableRng + RustToCuda, R: LineageReference<H> + DeviceCopy, S: LineageStore<H, R> + RustToCuda, X: EmigrationExit<H, G, R, S> + RustToCuda, D: DispersalSampler<H, G> + RustToCuda, C: CoalescenceSampler<H, R, S> + RustToCuda, T: TurnoverRate<H> + RustToCuda, N: SpeciationProbability<H> + RustToCuda, E: MinSpeciationTrackingEventSampler<H, G, R, S, X, D, C, T, N> + RustToCuda, I: ImmigrationEntry + RustToCuda, A: SingularActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> + RustToCuda, ReportSpeciation: Boolean, ReportDispersal: Boolean>(
) -> &'static CStr