Function rustcoalescence_algorithms_cuda::parallelisation::monolithic::simulate

pub fn simulate<'k, H: Habitat + RustToCuda, G: PrimeableRng + RustToCuda, R: LineageReference<H> + DeviceCopy, S: LineageStore<H, R> + RustToCuda, X: EmigrationExit<H, G, R, S> + RustToCuda, D: DispersalSampler<H, G> + RustToCuda, C: CoalescenceSampler<H, R, S> + RustToCuda, T: TurnoverRate<H> + RustToCuda, N: SpeciationProbability<H> + RustToCuda, E: MinSpeciationTrackingEventSampler<H, G, R, S, X, D, C, T, N> + RustToCuda, I: ImmigrationEntry + RustToCuda, A: SingularActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> + PeekableActiveLineageSampler<H, G, R, S, X, D, C, T, N, E, I> + RustToCuda, P: Reporter, L: LocalPartition<P>>(
    simulation: Simulation<H, G, R, S, X, D, C, T, N, E, I, A>, 
    kernel: &'k mut SimulationKernel<'k, H, G, R, S, X, D, C, T, N, E, I, A, <<WaterLevelReporterStrategy as WaterLevelReporterConstructor<'_, L::IsLive, P, L>>::WaterLevelReporter as Reporter>::ReportSpeciation, <<WaterLevelReporterStrategy as WaterLevelReporterConstructor<'_, L::IsLive, P, L>>::WaterLevelReporter as Reporter>::ReportDispersal>, 
    stream: &Stream, 
    config: (GridSize, BlockSize, DedupCache, NonZeroU64), 
    lineages: VecDeque<Lineage>, 
    event_slice: NonZeroUsize, 
    local_partition: &mut L
) -> Result<(NonNegativeF64, u64)>